[#162] Fix suggestion window retention, OCR threading, and enabled state persistence (#180)
* Address PR #162 review feedback: fix suggestion window retention, OCR threading, and enabled state persistence

- Retain AmbientSuggestionWindow in activeSuggestionWindow property to prevent premature deallocation (P1)
- Move VNImageRequestHandler.perform to background queue to avoid blocking main thread (P2)
- Only persist ambientAgentEnabled after start() succeeds to prevent stale enabled flag (P2)


* Persist max steps setting to UserDefaults and pass to session

- Back maxSteps slider with UserDefaults instead of ephemeral @State
- Read maxStepsPerSession from UserDefaults in startSession and pass to ComputerUseSession init


---------

+++ clients/macos/vellum-assistant/App/AppDelegate.swift
@@ -178,8 +178,8 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         }
 
         let provider = AnthropicProvider(apiKey: apiKey)
-        let storedMaxSteps = UserDefaults.standard.double(forKey: "maxStepsPerSession")
-        let maxSteps = storedMaxSteps > 0 ? Int(storedMaxSteps) : 50
+        let storedMaxSteps = UserDefaults.standard.integer(forKey: "maxStepsPerSession")
+        let maxSteps = storedMaxSteps > 0 ? storedMaxSteps : 50
         let session = ComputerUseSession(task: task, provider: provider, maxSteps: maxSteps)
         currentSession = session
